install `util` module so webpack can find it.
node.js loads the `util` module by default. webpack does not.

to fix, we first need to `yarn add util`

after that we need to add the following to `webpack.config.js`

```
resolve: {
    alias: {
        util: “util”
    }
}
```

this tells webpack to swap out any calls to the node.js’s built-in util library for the one we just installed.